# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Simva, AKA Simple Validator, AKA SurveyManager is the evolution of a tool created to integrate and ease the task of validating serious games using pre-post surveys as a control point. This version allows to create more complex studies with more variety of activities and different allocations.",
    "version": "1.0.0",
    "title": "Simva API",
    "contact": {
      "email": "ivanjper@ucm.es"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/groups/{id}/studies": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the assigned studies to the group",
        "operationId": "getGroupStudies",
        "description": "Obtains the list of studies assigned to the group\n",
        "responses": {
          "200": {
            "description": "list of studies assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/study"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Group not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/groupId"
          }
        ]
      }
    },
...
    "/studies": {
      "get": {
        "tags": [
          "admins",
          "teachers",
          "students"
        ],
        "summary": "gets the list of studies for the current teacher",
        "operationId": "getStudies",
        "description": "Obtains the list of studies for the current teacher.\n",
        "responses": {
          "200": {
            "description": "the list of studies for this teacher",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/study"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds an study for the current teacher",
        "operationId": "addStudy",
        "description": "Creates a new study for the current teacher.\n",
        "responses": {
          "200": {
            "description": "study created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/study"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/add_study_body"
              }
            }
          },
          "description": "Study to be created"
        }
      }
    },
    "/studies/import": {
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds an study for the current teacher",
        "operationId": "addStudyFromImport",
        "description": "Creates a new study for the current teacher.\n",
        "responses": {
          "200": {
            "description": "study created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/study"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/upload_file"
              }
            }
          },
          "description": "Study to be created"
        }
      }
    },
    "/studies/{studyid}": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the study with the given ID",
        "operationId": "getStudy",
        "description": "Obtains the requested study\n",
        "responses": {
          "200": {
            "description": "study object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/study"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      },
      "put": {
        "tags": [
          "teachers"
        ],
        "summary": "updates the study",
        "operationId": "updateStudy",
        "description": "Updates an existing stidy",
        "responses": {
          "200": {
            "description": "item created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Teacher does not own the class"
          },
          "404": {
            "description": "Requested study does not exist"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/study"
              }
            }
          },
          "description": "study to be modified"
        }
      },
      "delete": {
        "tags": [
          "teachers"
        ],
        "summary": "Deletes the study with the given ID",
        "operationId": "deleteStudy",
        "description": "Deleted the designed study\n",
        "responses": {
          "200": {
            "description": "Study deleted"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      }
    },
    "/studies/{studyid}/export": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the study export with the given ID",
        "operationId": "getStudyExport",
        "description": "Obtains the requested study Export\n",
        "responses": {
          "200": {
            "description": "study object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/study"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      }
    },
...
    "/studies/{studyid}/schedule": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "gets the list of scheduled activities for the student",
        "operationId": "getSchedule",
        "description": "Obtains the list of scheduled activities for the current\nstudent and study, and its completion status. Hides the\ncurrent used test to the user.\n",
        "responses": {
          "200": {
            "description": "The list of scheduled activities in order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "next": {
                      "type": "string",
                      "description": "The next activity ID"
                    }
                  },
                  "activities": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/activity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      }
    },
...
    "/studies/{studyid}/groups": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the assigned groups to the study",
        "operationId": "getStudyGroups",
        "description": "Obtains the list of groups assigned to the study\n",
        "responses": {
          "200": {
            "description": "list of groups assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/group"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Study not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      }
    },
    "/studies/{studyid}/tests": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the tests from a study",
        "operationId": "getStudyTests",
        "description": "Obtains the list of tests assigned to the study\n",
        "responses": {
          "200": {
            "description": "list of tests assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/test"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Study not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a test to the study",
        "operationId": "addTestToStudy",
        "description": "Adds a test for the current group\n",
        "responses": {
          "200": {
            "description": "test created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/test"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/test"
              }
            }
          },
          "required": true
        }
      }
    },
    "/studies/{studyid}/participants": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the list of participants for this study",
        "operationId": "getStudyParticipants",
        "description": "Obtains the list of participants of the study\n",
        "responses": {
          "200": {
            "description": "list of participants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "study not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      }
    },
    "/studies/{studyid}/tests/{testid}": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the test with the given ID",
        "operationId": "getTest",
        "description": "Obtains the requested test\n",
        "responses": {
          "200": {
            "description": "test object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/test"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      },
      "put": {
        "tags": [
          "teachers"
        ],
        "summary": "updates the test",
        "operationId": "updateTest",
        "description": "Updates an existing test",
        "responses": {
          "200": {
            "description": "item updated"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Teacher does not own the class"
          },
          "404": {
            "description": "Requested study does not exist"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/test"
              }
            }
          },
          "description": "test to be modified"
        }
      },
      "patch": {
        "summary": "Modifies the test.",
        "tags": [
          "teachers"
        ],
        "operationId": "patchTest",
        "description": "Modifies the test\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/test"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test updated",
            "schema": {
              "$ref": "#/components/schemas/test"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Teacher does not own the class"
          },
          "404": {
            "description": "Requested study does not exist"
          }
        }
      },
      "delete": {
        "tags": [
          "teachers"
        ],
        "summary": "deletes the test with the given ID",
        "operationId": "deleteTest",
        "description": "Deleted the selected test\n",
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/studies/{studyid}/tests/{testid}/activities": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the list of activities for the selected test",
        "operationId": "getTestActivities",
        "description": "Obtains the list of activities for the selected test\n",
        "responses": {
          "200": {
            "description": "the list of activities for this test",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/activity"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a Activity to a test",
        "operationId": "addActivityToTest",
        "description": "Creates a new activity for the current teacher.\n",
        "responses": {
          "200": {
            "description": "activity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activity"
              }
            }
          }
        }
      }
    },
    "/studies/{studyid}/allocator": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the allocator from a study",
        "operationId": "getStudyAllocator",
        "description": "Obtains the allocator used by the study\n",
        "responses": {
          "200": {
            "description": "the allocator object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/allocator"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Study not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ]
      },
      "put": {
        "tags": [
          "teachers"
        ],
        "summary": "set the allocator to the study",
        "operationId": "setStudyAllocator",
        "description": "Updates the allocator from the study\n",
        "responses": {
          "200": {
            "description": "updated correctly"
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/allocator"
              }
            }
          },
          "description": "study to be modified"
        }
      }
    },
...
  },
  "security": [
    {
      "OAuth2": [
        "read",
        "write"
      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/auth",
            "tokenUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/token",
            "scopes": {
              "read": "Grants read access",
              "write": "Grants write access"
            }
          }
        }
      }
    },
    "parameters": {
      "groupId": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string",
          "example": "5d07a0767c1405004bc702af"
        },
        "description": "The group ID"
      },
...
      "studyId": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string",
          "example": "5d07a0767c1405004bc702af"
        },
        "description": "The study ID"
      },
...
      "testId": {
        "in": "path",
        "name": "testid",
        "required": true,
        "schema": {
          "type": "string",
          "example": "5d07a0767c1405004bc702at"
        },
        "description": "The test ID"
      },
      "searchString": {
        "in": "query",
        "name": "searchString",
        "description": "pass an optional search string for result filtering",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "skip": {
        "in": "query",
        "name": "skip",
        "description": "number of records to skip for pagination",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0
        }
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "description": "maximum number of records to return",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "maximum": 50
        }
      },
...
    },
...
    "schemas": {
...
      "group": {
        "type": "object",
        "required": [
          "name",
          "owners",
          "participants",
          "created"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Test group for School"
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "link": {
            "type": "object",
            "x-mongoose-type": "Mixed"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "study": {
        "type": "object",
        "required": [
          "name",
          "tests",
          "allocator",
          "owners",
          "created"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Survey for SG validation"
          },
          "tests": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "allocator": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          },
          "active": {
            "type": "string",
            "format": "boolean",
            "example": true
          }
        }
      },
      "test": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Test A for my study"
          },
          "study": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "activities": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          }
        }
      },
...
      "upload_file": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "x-swagger-mongoose": {
          "exclude-schema": true
        }
      },
...
      "add_study_body": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Name of the Study"
          }
        },
        "x-swagger-mongoose": {
          "exclude-schema": true
        }
      },
...
    }
  }
}